Single Button Style for all buttons
All buttons use AdaptiveButtonStyle()
